Calculate actual number of turns

+++ n_brick.cpp
@@ -17,7 +17,7 @@ bool operator==(const n_brick& lhs, const n_brick& rhs) noexcept
   return lhs.get() == rhs.get();
 }
 
-n_brick operator*(const n_brick& w, const int n) noexcept
+n_brick operator*(const n_brick& w, const double n) noexcept
 {
   return n_brick(w.get() * n);
 }
