+++ n_ore.cpp
@@ -17,7 +17,7 @@ bool operator==(const n_ore& lhs, const n_ore& rhs) noexcept
   return lhs.get() == rhs.get();
 }
 
-n_ore operator*(const n_ore& w, const int n) noexcept
+n_ore operator*(const n_ore& w, const double n) noexcept
 {
   return n_ore(w.get() * n);
 }
