+++ n_wheat.cpp
@@ -17,7 +17,7 @@ bool operator==(const n_wheat& lhs, const n_wheat& rhs) noexcept
   return lhs.get() == rhs.get();
 }
 
-n_wheat operator*(const n_wheat& w, const int n) noexcept
+n_wheat operator*(const n_wheat& w, const double n) noexcept
 {
   return n_wheat(w.get() * n);
 }
