+++ n_wool.cpp
@@ -17,7 +17,7 @@ bool operator==(const n_wool& lhs, const n_wool& rhs) noexcept
   return lhs.get() == rhs.get();
 }
 
-n_wool operator*(const n_wool& w, const int n) noexcept
+n_wool operator*(const n_wool& w, const double n) noexcept
 {
   return n_wool(w.get() * n);
 }
